Configures Redis for Upstash integration
Migrates the Redis configuration to be compatible with Upstash,
using `redis` package instead of `ioredis`. This includes changes to
use `REDIS_URL` and `REDIS_TOKEN` for authentication, as required by
Upstash, and updates all relevant methods to use the new client.

A new markdown file is added with instructions to set up Upstash.